Split `Init`, `Model`, `Update` and `View` logic
