+++ tui/stopwatch/model.go
@@ -1,0 +1,31 @@
+package stopwatch
+
+import (
+	"time"
+
+	"github.com/charmbracelet/bubbles/help"
+)
+
+// MODEL
+// -----
+
+// Represents a stopwatch
+type stopwatchModel struct {
+	running bool            // Whether the stopwatch is running
+	start   time.Time       // The time the stopwatch was started
+	end     time.Time       // The time the stopwatch was stopped
+	elapsed time.Duration   // The elapsed time
+	laps    []time.Duration // The lap times
+
+	keys KeyMap     // Key bindings for the stopwatch view
+	help help.Model // Help menu model
+}
+
+// Creates a new stopwatch model. Initializes the default key bindings and
+// help menu. The stopwatch is "stopped" by default.
+func newStopwatchModel() *stopwatchModel {
+	return &stopwatchModel{
+		keys: DefaultKeyMap,
+		help: help.New(),
+	}
+}
