+++ tui/stopwatch/update.go
@@ -1,0 +1,93 @@
+package stopwatch
+
+import (
+	"time"
+
+	"github.com/Shresht7/go-time/tui/helpers"
+	"github.com/charmbracelet/bubbles/key"
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+// UPDATE
+// ------
+
+// The Update function is called when events happen, like keypresses
+func (m *stopwatchModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+
+	// Window Resize Event
+	case tea.WindowSizeMsg:
+		// If we set a width on the help menu it can gracefully truncate
+		// its view as needed.
+		m.help.Width = msg.Width
+
+	// Stopwatch Start
+	case msgStart:
+		m.start = msg.t
+		m.running = true
+		m.keys.Space.SetHelp("<spacebar>", "stop")
+		return m, nil
+
+	// Stopwatch Stop
+	case msgStop:
+		m.end = time.Now()
+		m.running = false
+		m.keys.Space.SetHelp("<spacebar>", "start")
+		return m, nil
+
+	// Stopwatch Resume
+	case msgResume:
+		m.start = time.Now().Add(-m.elapsed)
+		m.running = true
+		m.keys.Space.SetHelp("<spacebar>", "stop")
+		return m, nil
+
+	// Stopwatch Lap
+	case msgLap:
+		m.laps = append(m.laps, m.elapsed)
+		return m, nil
+
+	// Stopwatch Reset
+	case msgReset:
+		m.elapsed = 0
+		m.running = false
+		m.keys.Space.SetHelp("<spacebar>", "start")
+		return m, nil
+
+	// Stopwatch Tick
+	case helpers.MsgTick:
+		if m.running {
+			m.elapsed = time.Since(m.start)
+		}
+		return m, tick()
+
+	// Key Press
+	case tea.KeyMsg:
+		switch {
+
+		// Quit
+		case key.Matches(msg, DefaultKeyMap.Quit):
+			return m, tea.Quit
+
+		// Space
+		case key.Matches(msg, DefaultKeyMap.Space):
+			if m.running {
+				return m, m.Stop
+			} else {
+				return m, m.Resume
+			}
+
+		// R
+		case key.Matches(msg, DefaultKeyMap.R):
+			return m, m.Reset
+
+		// Enter
+		case key.Matches(msg, DefaultKeyMap.Enter):
+			return m, m.Lap
+
+		}
+
+	}
+
+	return m, nil
+}
